requirements.txt
streamlit==1.28.0
pandas==2.1.1
plotly==5.17.0
requests==2.31.0
python-dotenv==1.0.0
PyYAML==6.0.1